+++ README.md
@@ -32,7 +32,7 @@
 
 ### 요구사항
 
-- [ ] 본인만의 기준으로 일관된 코드를 작성해주세요.
+- [x] 본인만의 기준으로 일관된 코드를 작성해주세요.
 - [x] 기능 단위로 나누어 커밋을 해주세요.
 - [x] 재사용 가능성이 높은 컴포넌트, Hooks가 보이면 재사용성을 고려하여 만들어요.
 - [x] 별도의 From, Validation 라이브러리를 사용하지 않고 useState로만 구현해요.
